+++ NEWS.md
@@ -1,4 +1,4 @@
-xpmdf 0.1.0 (development)
+xmpdf 0.1.0 (development)
 =========================
 
 Initial features
@@ -16,7 +16,7 @@ Initial features
   2. `set_bookmarks_pdftk()` which wraps `pdftk` command-line tool
 
 * `docinfo()` provides an `{R6}` object to represent pdf documentation info entries.
-* `get_docinfo()` reads in pdf documentation info entries from a pdf file.
+* `get_docinfo()` reads in pdf documentation info entries from pdf file(s).
   `get_docinfo()` will try to use the following helper functions in the following order:
 
   1. `get_docinfo_pdftools()` which wraps `pdftools::pdf_info()`
@@ -27,6 +27,16 @@ Initial features
 
   1. `set_docinfo_gs()` which wraps `ghostscript` command-line tool
   2. `set_docinfo_pdftk()` which wraps `pdftk` command-line tool
+
+* `get_xmp()` reads in xmp metadata from media file(s).
+  `get_xmp()` will try to use the following helper functions in the following order:
+
+  1. `get_xmp_exiftool()` which wraps `exiftool` command-line tool
+
+* `set_xmp()` sets xmp metadata in media file(s).
+  `set_xmp()` will try to use the following helper functions in the following order:
+
+  1. `set_xmp_exiftool()` which wraps `exiftool` command-line tool
 
 * `n_pages()` returns the number of pages in the (pdf) file(s).
   `n_pages()` will try to use the following helper functions in the following order:
@@ -42,6 +52,8 @@ Initial features
   * `supports_set_bookmarks()`
   * `supports_get_docinfo()`
   * `supports_set_docinfo()`
+  * `supports_get_xmp()`
+  * `supports_set_xmp()`
   * `supports_n_pages()`
 
 * The following functions detect support for various command-line tools needed for some